last endpoint check

+++ lib/endpoints.rb
@@ -17,6 +17,7 @@ class LeoTamer
 
     delete "/delete_endpoint" do
       endpoint = required_params(:endpoint)
+      raise "You can't delete the last endpoint" if @@manager.get_endpoints.size == 1
       @@manager.delete_endpoint(endpoint)
       200
     end
